Create Trebuchet?! Part 1
Unrefined Java code that scans in the calibration doc, grabbing the first and last numerals for each line to translate it for the trebuchet. 

Created quickly in Main, but would function better as its own calibration method, where you pass a file, it does all of its processing, then returns the translation.